Adding on branch-121 as well as on main trunk
Module that combines exonerate/e2g predictions with genewise predictions
on a virtual contig to produce genes with UTRs.  This approach will
eventually replace the combining code in TargettedGeneE2G and allow us
to get rid of some of the vile SwissProt parsing scripts currently used
in the genebuild. Woohoo!