new SeqFetcher, wrapper of the SeqFetcher Bio::DB::Flat::OBDAIndex (in bioperl main trunk) to be used with Steve indexed databases.
You'll need to check out bioperl-main-trunk to be able to use it.

It works like any other SeqFetcher:

my $seqfetcher = Bio::EnsEMBL::Pipeline::SeqFetcher::OBDAIndexSeqFetcher->new(
										-db     => $db,
										-format => $format,
										);

where $db is an arrayref of databases.
For each database, Steve's indexing system produces from a file, e.g. '/path/sequences.fa',
and a directory '/path/sequences/' with two files, where in one of them the file
'sequences.fa' is specified

For the OBDAIndexSeqFetche, each element of the array
of databases is the index-directory name: '/path/sequences'

The format is optional and is defaulted to "FASTA"

As usual, the sequences can be retrieved as

my $seq = $seqfetcher->get_Seq_by_acc($acc);