Runnable/RunnableDB for the Genoscope Exofish procedure.
The runnable is simply a wrapper for wutblastx, adding the correct matrix
and parameters, and filtering the hits according to percentage identity
and length (using Figure 1 in the Exofish paper).

To obtain Genoscope-style "Ecores", the dna_align_features that are
written by the RunnableDB need to be projected onto the genomic sequence
in a strand-independent fashion, with overlapping regions being merged.
This is currently done by a downstream script, because the pre-Ecore
features generated here are useful in their own right.