temporarily changed to proto=>6 based on what Mario found: Dear isg,
 I think there is problem in the

 IO::Socket::INET

 module installed in the rlx nodes.

 The method call

 $server = IO::Socket::INET->new(
				PeerAddr => $host,
				PeerPort => $port,
				Proto    => 'tcp',
				Type     => SOCK_STREAM,
				Timeout  => 10,
			       );

fails in the rlx nodes complaining about the string 'tcp'.


In the rlx distribution of INET.pm we have the lines

if(defined $proto  && $proto =~ /\D/) {
    if(@proto = getprotobyname($proto)) {
      $proto = $proto[2] || undef;
    }
    else {
      $@ = "Bad protocol '$proto'";
      return;
    }
  }

whereas in the other instalations is

if(defined $proto) {
    if (@proto = ( $proto =~ m,\D,
		? getprotobyname($proto)
		: getprotobynumber($proto))
    ) {
	$proto = $proto[2] || undef;
    }
    else {
	$@ = "Bad protocol '$proto'";
      return;
    }
}

conclusion: the rlx version only accepts numbers when the current
documentation for the modules tells that a protocol name like 'tcp' can
be used.

 Thanks,

Mario