Fixed bug that was setting incorrect phases for UTR-containing transcripts that have a translation completely inside a single exon.
NOTE: This change can give rise to transcripts that violate the
$exons[$i]->phase = $exons[$i-1]->end_phase convention. In particular,
if the beginning of translation is flush with the start of an exon,
and there are additional 5' non-translating exons, then these all have
phase = end_phase = -1, whereas the phase of the first coding exon is 0.
After discussion, it was decided that this is the "right" thing to do.