my @classification = split(",",$node->classification(",", 1));
When fetching the node classification, pass both the separator (",") and also the full flag (1), so that the full classification is fetched.
This is required for reptiles and birds as we need 'Sauropsida' to be their shared taxonomic node. 'Sauropsida' is not inserted into the meta table if we do not load the full classification.