(1) Added use Bio::Seq or else the seq subroutine would not be found when called inside the trim_if_possible subroutine.
(2) When high_quality_sequence_start and high_quality_sequence_end positions
are the same for a given cDNA or EST, return the original seq object unchanged
instead of "undef". An undefined object will not be accepted in the second
part of the script where the subroutine clip_if_ncessary is called to clip
off polyAs, and will cause the script to die.